Create client base library, and refactor others.
The client_base library is used in Apollo and SaturnV for common
utilities. Other libraries were changed to match the new naming
convention that can be found in the wiki.

Add PathResolver to query the OS for preferred locations to store
certain runtime data. Currently, only the OSX host implementation
exists.