Semi-major rearchitecture for Buffer.
All write(...) functions from Mapper have now been moved to
TypedBuffer<> for simplication reasons. Mapper is now known as
ChannelMapper and simply provides references to the proper locations
in the buffer for each mapped channel. Work has begun to support
entire buffer read-ins now.

Note: This is a broken build. It is a nightly commit.